Get ACE, core and coupled tests passing on MPS (#1672)
I fixed/skipped all relevant tests in `fme/core`, `fme/ace`, and
`fme/coupled` when calling `FME_USE_MPS=1 pytest`.
I didn't touch `fme/downscaling` since there is a lot in flux there
right now.

+++ fme/ace/inference/test_segmented.py
@@ -105,7 +105,8 @@ def test_inference_segmented_entrypoint():
         initial_condition = xr.Dataset(
             {
                 "prog": xr.DataArray(
-                    np.random.rand(2, 16, 32), dims=["sample", "lat", "lon"]
+                    np.random.rand(2, 16, 32).astype(np.float32),
+                    dims=["sample", "lat", "lon"],
                 )
             }
         )
